Moved the base nudge button class out of the select tool
